+++ packages/cli/src/scripts/create-wrappers/main.ts
@@ -67,6 +67,7 @@ export default async function createWrappers(packageName: string, outDir: string
         .setIsAbstract(declaration._ui5abstract ?? false)
         .setSince(declaration._ui5since)
         .setIsExperimental(declaration._ui5experimental)
+        .setIsDeprecated(declaration.deprecated)
     );
     wrapper.addRenderer(new ExportsRenderer());