+++ packages/main/src/components/MessageView/MessageItem.tsx
@@ -1,22 +1,26 @@
 'use client';
 
+import IconMode from '@ui5/webcomponents/dist/types/IconMode.js';
 import ListItemType from '@ui5/webcomponents/dist/types/ListItemType.js';
+import WrappingType from '@ui5/webcomponents/dist/types/WrappingType.js';
 import ValueState from '@ui5/webcomponents-base/dist/types/ValueState.js';
 import iconArrowRight from '@ui5/webcomponents-icons/dist/slim-arrow-right.js';
-import { useStylesheet } from '@ui5/webcomponents-react-base';
+import { useI18nBundle, useStylesheet } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
 import type { ReactNode } from 'react';
-import { Children, forwardRef, useContext, useEffect, useRef, useState } from 'react';
+import { Children, isValidElement, forwardRef, useContext, useEffect, useRef, useState } from 'react';
 import { FlexBoxAlignItems, FlexBoxDirection } from '../../enums/index.js';
+import { COUNTER, HAS_DETAILS } from '../../i18n/i18n-defaults.js';
 import { MessageViewContext } from '../../internal/MessageViewContext.js';
 import type { CommonProps } from '../../types/index.js';
 import { Icon } from '../../webComponents/Icon/index.js';
 import { Label } from '../../webComponents/Label/index.js';
+import type { LinkPropTypes } from '../../webComponents/Link/index.js';
 import type { ListItemCustomDomRef, ListItemCustomPropTypes } from '../../webComponents/ListItemCustom/index.js';
 import { ListItemCustom } from '../../webComponents/ListItemCustom/index.js';
 import { FlexBox } from '../FlexBox/index.js';
 import { classNames, styleData } from './MessageItem.module.css.js';
-import { getIconNameForType } from './utils.js';
+import { getIconNameForType, getValueStateMap } from './utils.js';
 
 export interface MessageItemPropTypes extends Pick<ListItemCustomPropTypes, 'accessibleName' | 'tooltip'>, CommonProps {
   /**
@@ -60,8 +64,10 @@ export interface MessageItemPropTypes extends Pick<ListItemCustomPropTypes, 'acc
 const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((props, ref) => {
   const { titleText, subtitleText, counter, type = ValueState.Negative, children, className, ...rest } = props;
   const [isTitleTextOverflowing, setIsTitleTextIsOverflowing] = useState(false);
+  const [titleTextStr, setTitleTextStr] = useState('');
   const titleTextRef = useRef<HTMLSpanElement>(null);
   const hasDetails = !!(children || isTitleTextOverflowing);
+  const i18nBundle = useI18nBundle('@ui5/webcomponents-react');
 
   useStylesheet(styleData, MessageItem.displayName);
 
@@ -78,14 +84,14 @@ const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((prop
 
   const handleListItemClick = (e) => {
     if (hasDetails) {
-      selectMessage(props);
+      selectMessage({ ...props, titleTextStr });
       if (typeof rest.onClick === 'function') {
         rest.onClick(e);
       }
     }
   };
 
-  const handleKeyDown = (e) => {
+  const handleKeyDown: ListItemCustomPropTypes['onKeyDown'] = (e) => {
     if (typeof rest.onKeyDown === 'function') {
       rest.onKeyDown(e);
     }
@@ -108,7 +114,6 @@ const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((prop
           isChildOverflowing = firstChild.scrollWidth > firstChild.clientWidth;
         }
       }
-
       setIsTitleTextIsOverflowing(isTargetOverflowing || isChildOverflowing);
     });
     if (!hasChildren && titleTextRef.current) {
@@ -119,11 +124,20 @@ const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((prop
     };
   }, [hasChildren]);
 
+  useEffect(() => {
+    if (typeof titleText === 'string') {
+      setTitleTextStr(titleText);
+    } else if (isValidElement(titleText) && typeof (titleText.props as LinkPropTypes)?.children === 'string') {
+      // @ts-expect-error: props.children is available and a string
+      setTitleTextStr(titleText.props.children);
+    }
+  }, [titleText]);
+
   return (
     <ListItemCustom
       onClick={handleListItemClick}
       onKeyDown={handleKeyDown}
-      data-title={titleText}
+      data-title={titleTextStr}
       data-type={type}
       type={hasDetails ? ListItemType.Active : ListItemType.Inactive}
       {...rest}
@@ -132,7 +146,7 @@ const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((prop
     >
       <FlexBox alignItems={FlexBoxAlignItems.Center} className={messageClasses}>
         <div className={classNames.iconContainer}>
-          <Icon name={getIconNameForType(type as ValueState)} className={classNames.icon} />
+          <Icon name={getIconNameForType(type as ValueState)} className={classNames.icon} mode={IconMode.Decorative} />
         </div>
         <FlexBox
           direction={FlexBoxDirection.Column}
@@ -143,10 +157,22 @@ const MessageItem = forwardRef<ListItemCustomDomRef, MessageItemPropTypes>((prop
               {titleText}
             </span>
           )}
-          {titleText && subtitleText && <Label className={classNames.subtitle}>{subtitleText}</Label>}
+          {titleText && subtitleText && (
+            <Label className={classNames.subtitle} wrappingType={WrappingType.None}>
+              {subtitleText}
+            </Label>
+          )}
         </FlexBox>
-        {counter != null && <span className={classNames.counter}>{counter}</span>}
-        {hasDetails && <Icon className={classNames.navigation} name={iconArrowRight} />}
+        {counter != null && (
+          <span className={classNames.counter} aria-label={`. ${i18nBundle.getText(COUNTER)} ${counter}`}>
+            {counter}
+          </span>
+        )}
+        {hasDetails && <Icon className={classNames.navigation} name={iconArrowRight} mode={IconMode.Decorative} />}
+        {hasDetails && <span className={classNames.pseudoInvisibleText}>. {i18nBundle.getText(HAS_DETAILS)}</span>}
+        {type !== ValueState.None && type !== ValueState.Information && (
+          <span className={classNames.pseudoInvisibleText}>. {getValueStateMap(i18nBundle)[type]}</span>
+        )}
       </FlexBox>
     </ListItemCustom>
   );